flask/flask_dash_web/requirements.txt
requests
flask
dash
gunicorn
pandas
dash-bootstrap-components
psycopg2-binary